duc::types

Struct DucLocalState

pub struct DucLocalState {

    pub scope: String,
    pub scroll_x: f64,
    pub scroll_y: f64,
    pub zoom: f64,
    pub is_binding_enabled: bool,
    pub current_item_stroke: Option<ElementStroke>,
    pub current_item_background: Option<ElementBackground>,
    pub current_item_opacity: f32,
    pub current_item_font_family: String,
    pub current_item_font_size: f64,
    pub current_item_text_align: TEXT_ALIGN,
    pub current_item_start_line_head: Option<DucHead>,
    pub current_item_end_line_head: Option<DucHead>,
    pub current_item_roundness: f64,
    pub pen_mode: bool,
    pub view_mode_enabled: bool,
    pub objects_snap_mode_enabled: bool,
    pub grid_mode_enabled: bool,
    pub outline_mode_enabled: bool,
    pub manual_save_mode: bool,
    pub decimal_places: i32,

}

Fields
scope: String

The current scope of the design. mm, cm, m, in, ft, yd, mi, etc…

scroll_x: f64scroll_y: f64zoom: f64is_binding_enabled: boolcurrent_item_stroke: Option<ElementStroke>

Current item is usually a quick access state to apply as default to certain things when drawing

current_item_background: Option<ElementBackground>current_item_opacity: f32current_item_font_family: Stringcurrent_item_font_size: f64current_item_text_align: TEXT_ALIGNcurrent_item_start_line_head: Option<DucHead>current_item_end_line_head: Option<DucHead>current_item_roundness: f64pen_mode: bool

Pen mode is enabled, creates a better experience for drawing with a pen

view_mode_enabled: bool

In view mode the user is not allowed to edit the canvas.

objects_snap_mode_enabled: bool

Object snapping on the environment is enabled

grid_mode_enabled: bool

Available grids are visible

outline_mode_enabled: bool

Whether to disable the fill on all shapes

manual_save_mode: bool

When enabled, the version graph is not updated automatically. The user needs to manually update the graph for new versions to be saved in version control.

decimal_places: i32